moved validators into common
