+++ lib/src/test/java/com/thenewboston/api/bankapi/repository/BankRepositoryTest.kt
@@ -15,7 +15,7 @@ import com.thenewboston.data.dto.bankapi.crawl.response.Crawl
 import com.thenewboston.data.dto.bankapi.invalidblockdto.InvalidBlock
 import com.thenewboston.data.dto.bankapi.validatorconfirmationservicesdto.ConfirmationServices
 import com.thenewboston.data.dto.bankapi.validatorconfirmationservicesdto.ConfirmationServicesList
-import com.thenewboston.data.dto.bankapi.validatordto.ValidatorList
+import com.thenewboston.data.dto.common.response.ValidatorList
 import com.thenewboston.utils.Mocks
 import com.thenewboston.utils.PaginationOptions
 import com.thenewboston.utils.Some
